Refactor Warning component to use StatusPage component
**Why**: Avoid duplication

[skip changelog]
